+++ cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "sspšg-discord-bot"
+version = "0.1.1"
+authors = ["František Vojáček <vojacek.fr.2021@ssps.cz>"]
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[[bin]]
+name = "bot"
+path = "kod/bot.rs"
+
+[dependencies]
+log = "0.4"
+env_logger = "0.10"
+dotenv = "0.15"
+serenity = { version = "0.11", default-features = false, features = ["client", "gateway", "rustls_backend", "model"] }
+tokio = { version = "1.21.2", features = ["macros", "rt-multi-thread"] }
