+++ readme.md
@@ -1,0 +1,25 @@
+<div align="center">
+    <img src="design/banner/sspsg discord bot.svg" alt="SSPŠG Discord Bot"/>
+</div>
+<div align="center">
+    <img src="design/banner/sspsg discord.svg" alt="SSPŠG × Discord"/>
+</div>
+
+<br/>
+<img src="design/nadpis/funkce.svg" alt="Funkce">
+
+- [ ] Hlášení
+- [ ] Suplování
+    - [ ] Příkaz
+    - [ ] Interval
+    - [ ] Detekce změny
+- [ ] Služba
+    - [ ] Příkaz
+    - [ ] Interval
+
+<br/>
+<p align="center">
+    <a href="https://discord.com/api/oauth2/authorize?client_id=1080575281091326063&permissions=274877908992&scope=bot%20applications.commands">
+        <img src="design/banner/pridat.svg" alt="Přidat"/>
+    </a>
+</p>
